Day7: just use isize

+++ day7/src/main.rs
@@ -6,7 +6,7 @@ fn main() {
     println!("Part 2: {}", part2(&input))
 }
 
-fn part1(input: &Vec<usize>) -> usize {
+fn part1(input: &Vec<isize>) -> isize {
     input
         .iter()
         .map(|&x| fuel_spend_p1(input, x))
@@ -14,14 +14,14 @@ fn part1(input: &Vec<usize>) -> usize {
         .unwrap()
 }
 
-fn part2(input: &Vec<usize>) -> usize {
+fn part2(input: &Vec<isize>) -> isize {
     let &min = input.iter().min().unwrap();
     let &max = input.iter().max().unwrap();
 
     (min..=max).map(|x| fuel_spend_p2(input, x)).min().unwrap()
 }
 
-fn parse_input(s: &str) -> Result<Vec<usize>, ParseIntError> {
+fn parse_input(s: &str) -> Result<Vec<isize>, ParseIntError> {
     s.trim().split(',').map(str::parse).collect()
 }
 
@@ -29,22 +29,22 @@ fn input() -> &'static str {
     include_str!("../../input/day7.txt")
 }
 
-fn fuel_spend_p1(source_positions: &Vec<usize>, dest: usize) -> usize {
+fn fuel_spend_p1(source_positions: &Vec<isize>, dest: isize) -> isize {
     source_positions
         .iter()
-        .map(|&x| (dest as isize - x as isize).abs() as usize)
+        .map(|&x| (dest - x).abs())
         .sum()
 }
 
-fn fuel_spend_p2(source_positions: &Vec<usize>, dest: usize) -> usize {
+fn fuel_spend_p2(source_positions: &Vec<isize>, dest: isize) -> isize {
     source_positions
         .iter()
         .map(|&x| fuel_required_p2(x, dest))
         .sum()
 }
 
-fn fuel_required_p2(from: usize, to: usize) -> usize {
-    let distance = (from as isize - to as isize).abs() as usize;
+fn fuel_required_p2(from: isize, to: isize) -> isize {
+    let distance = (from as isize - to as isize).abs() as isize;
 
     (distance * (distance + 1)) / 2
 }
